ECS Fixes and Initialize render system
Renamed System Init to System Register.
Changed FSystemInit booleans to bitmask flags
Fixed a bug on System Type ID
Started work on Render System's init
Added Render System init to main

+++ ProjectCreation/ProjectCreation/Engine/GEngine.cpp
@@ -51,7 +51,7 @@ SystemManager* GEngine::GetSystemManager()
 
 float GEngine::GetDeltaTime()
 {
-        return m_XTime.Delta();
+        return (float)m_XTime.Delta();
 }
 
 double GEngine::GetTotalTime()
